Unit tests: re-organises directories
