+++ tests/unit/specs/layouts/auth.spec.js
